Move recent-*-stores to file-system/controllers
